test: isolate image-history regressions from sibling fixes
